feat(back): implement RF09 stats registration logic

+++ src/main/java/com/doback/E_rank/infrastructure/repository/jpa/UsuariosJpa.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UsuariosJpa extends JpaRepository<UsuariosModel, Long> {
+public interface UsuariosJpa extends JpaRepository<UsuariosModel, Integer> {
     Optional<UsuariosModel> findByEmail(String email);
     List<UsuariosModel> findByNicknameContainingIgnoreCase(String nickname);
 }
